updated ScreenMenu.js
reallocate all files
