+++ src/components/Menus/ScreenMenu.js
@@ -2,18 +2,21 @@ import React, { useContext } from "react";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { AuthContext } from "../context/authContext";
 // Import Screens - Fix any case sensitivity issues here
-import SplashScreen from '../screens/SplashScreen';
-import WelcomePage from '../screens/WelcomePage';
-import LoginPage from '../screens/Loginpage'; // Fixed case
-import Registration from '../screens/registration/Registration';
-import HomeScreen from '../screens/HomeScreen';
-import ProfilePage from '../screens/profile/Profile';
-import EditProfile from '../screens/profile/EditProfile/EditProfile';
-import EditAbout from '../screens/EditAbout';
-import EditLookingFor from '../screens/EditLookingFor';
-import EditGames from '../screens/EditGames';
-import EditPackage from '../screens/EditPackage';
+import SplashScreen from '../Screens/SplashScreen';
+import WelcomePage from '../Screens/WelcomePage';
+import LoginPage from '../Screens/Loginpage'; // Fixed case
+import Registration from '../Screens/registration/Registration';
+import HomeScreen from '../Screens/HomeScreen';
+import ProfilePage from '../Screens/profile/Profile';
+import EditProfile from '../Screens/profile/EditProfile/EditProfile';
+import EditAbout from '../Screens/profile/EditProfile/EditAbout';
+import EditLookingFor from '../Screens/profile/EditProfile/EditLookingFor';
+import EditGames from '../Screens/profile/EditProfile/EditGames';
+import EditPackage from '../Screens/profile/EditProfile/EditPackage';
 import HeaderMenu from '../Menus/HeaderMenu';
+import Chat from '../Screens/Chat';
+import Explore from '../Screens/Explore';
+import Settings from '../Screens/Settings';
 
 const Stack = createNativeStackNavigator();
 
@@ -28,10 +31,23 @@ const ScreenMenu = () => {
           <Stack.Screen
             name="Home"
             component={HomeScreen}
-            options={{
-              title: "Full Stack App",
-              headerRight: () => <HeaderMenu />,
-            }}
+              options={{ headerShown: false }}
+          />
+          <Stack.Screen
+            name="Chat"
+            component={Chat}
+              options={{ headerShown: false }}
+          />
+
+          <Stack.Screen
+            name="Explore"
+            component={Explore}
+              options={{ headerShown: false }}
+          />
+          <Stack.Screen
+            name="Settings"
+            component={Settings}
+              options={{ headerShown: false }}
           />
 
           <Stack.Screen
@@ -43,7 +59,7 @@ const ScreenMenu = () => {
             }}
           />
 
-<Stack.Screen
+          <Stack.Screen
             name="EditLookingFor"
             component={EditLookingFor}
             options={{
@@ -52,7 +68,7 @@ const ScreenMenu = () => {
             }}
           />
 
-<Stack.Screen
+          <Stack.Screen
             name="EditGames"
             component={EditGames}
             options={{
@@ -61,7 +77,7 @@ const ScreenMenu = () => {
             }}
           />
 
-<Stack.Screen
+          <Stack.Screen
             name="EditPackage"
             component={EditPackage}
             options={{
