+++ src/components/context/authContext.js
@@ -32,7 +32,7 @@ const AuthProvider = ({ children }) => {
   //default axios setting
   axios.defaults.headers.common["Authorization"] = `Bearer ${token}`;
   axios.defaults.baseURL =
-    "http://192.168.0.102:8080/api/v1";
+    "http://192.168.0.106:8080/api/v1";
 
   return (
     <AuthContext.Provider value={[state, setState]}>
